requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
apscheduler==3.10.4
sqlalchemy==2.0.35
httpx==0.27.2
feedparser==6.0.11
python-dotenv==1.0.1
google-generativeai==0.8.3
groq==0.11.0
anthropic==0.34.2
pydantic==2.9.2
alembic==1.13.3

# === API Framework ===
fastapi==0.115.0
uvicorn==0.30.6

# === Base de données ===
sqlalchemy==2.0.35
alembic==1.13.3

# === Scheduler ===
apscheduler==3.10.4

# === HTTP Client (async) ===
httpx==0.27.2

# === Flux RSS ===
feedparser==6.0.11

# === Variables d'environnement ===
python-dotenv==1.0.1

# === Validation des données ===
pydantic==2.9.2
pydantic-settings==2.5.2

# === IA — Gemini (gratuit par défaut) ===
google-generativeai==0.8.3

# === IA — Groq (gratuit alternatif) ===
groq==0.11.0

# === IA — Claude (optionnel) ===
anthropic==0.34.2